[roll] Roll fuchsia [minfs] Remove unused build arg.
The minfs_board_maximum_data_bytes was used to tell appmgr the maximum
size of the partition which it would use in cache clearing logic. This
was a temporary workaround for minfs not accounting for the partition
limit in its size calculations.

The minfs size calculations have been corrected and this value is no
longer used.

Original-Bug: 85718
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/606086
Original-Revision: 63f35abcb35b15a365e2f7ed39a941b6495a482e
GitOrigin-RevId: e6412eec9bee7a036dd1334d3c0eb3d29520c7ff
Change-Id: I218a4efc7374aaa9cde272f118c05cb02e29f181