[roll] Roll fuchsia [rust-3p-crates] num's `use_addcarry` only works on x86/x64
It appears that when num got updated, we enabled the `use_addcarry`
feature, but didn't end up re-generating the BUILD.gn. When re-run, this
will fail to compile on aarch64, because the `use_addcarry` feature only
works on x86 or x86_64.

This patch changes the configuration to only set this feature if we are
running on the appropriate target.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/607188
Original-Revision: ce482e2eb79ca69c255dfb7a12f6c293bf808f37
GitOrigin-RevId: 1205ec12c200b79926370422fb66b711e40003c2
Change-Id: I832020eab5c65af1f87c3544fac44e3cb1865451