[roll] Roll fuchsia [gidl] GIDL test for bytes/handles exceeding payload size
This replaces an existing Go test with a GIDL test. This test fails in
Rust and Dart but no attempt was made to fix the issues in this CL
because they require separate fixes to each of the bindings.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/606701
Original-Revision: e0810ef44d253b099c0b929265a99a05d4b1670e
GitOrigin-RevId: 1ce5983907ba4619f5bf25d6cde0b17cd1dcee35
Change-Id: Ib021eb1c02477b405b0541a76fee3e29928fe371