[roll] Roll fuchsia [network] Make UpdateResult generic over properties type
Parameterize UpdateResult on the interface properties type, which could
be (for example) a reference into the maintained interface state, an
owned type, or a key into the state.

This enables clients to more easily share results of updates to
interface state.

Original-Fixed: 87630
Test: fx test fidl-fuchsia-net-interfaces-ext-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601903
Original-Revision: b99962cd050811781e7cd805dcbae25ef74f4017
GitOrigin-RevId: f5cd21b4615c5486a4b9211f0aebe1ad55ee651c
Change-Id: Id648b1cb55fe47ac84f4a8114939b0d711dadca7