[roll] Roll fuchsia [netcfg] Impl fuchsia.net.virtualization/Control
Allow clients to create bridged networks by keeping track of interfaces
that can provide upstream connectivity and bridging them with interfaces
that are added to the virtual network by the client.

Due to the limitations of the current bridging API, we add or remove
interfaces from the bridge by destroying the bridge and rebuilding it.

Original-Fixed: 85078
Test: fx test netcfg-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/595149
Original-Revision: 2eb36f1f363047ecd54349da8888969d88a0ad72
GitOrigin-RevId: f8805a94c73bbfe768284ec88fa90257f2a88c1c
Change-Id: Ic6b555df973288019720b987ad6e5989f1075bc9