[roll] Roll fuchsia [diagnostics-data] Remove size_bytes in LogsData
This was previously used for computing budgets, but since we are no
longer tracking LogsData inside the archivist except for the response we
don't really need to store the size of the original message in the
response. This doesn't seem like useful information for clients.

Original-Bug: 85755
Tested: existing tests + backwards compatibility test in diagnostics-data

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585344
Original-Revision: 6229538ba574baa6ffe6a6f240dfbaa65bbfecde
GitOrigin-RevId: 96b8eebf4386b815e09be2ee2375bbe3f52edd85
Change-Id: Ic03bba503c60c652c78ced05a107a7aff600f44b