[roll] Roll fuchsia [ffx] ignore mdns packets with dubious addresses
When QEMU/AEMU are run using SLIRP they will write multicast packets out
to the host network approximately "as-is", however, those responses do
not necessarily indicate a viable return path. To date the code had
assumed that the host networking environment is sound with respect to
mdns propagation, but as it appears to not be so packets are now
validated against their source address.  If the packet does not contain
an answer that matches the source address it is now ignored, effectively
filtering out the NAT'd SLIRP targets. This may filter out desired
packets in some as yet not understood scenarios, in which case this
heuristic would need to be revised.

Original-Fixed: 88587
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/604646
Original-Revision: 98a2283e196555cfd80337c225783fbaaa854a61
GitOrigin-RevId: 44b52b19ad26fdfac5fe3bc16b5719c294616eac
Change-Id: I3dec6c82613599a6ce559db43f2921d67ac30ab7