[roll] Roll fuchsia [brcmfmac] Replace scan_request with boolean scan_in_progress field
The pointer copied into the scan_request field of struct
brcmf_cfg80211_info is not owned by brcmfmac, so a proper copy of the
data it contains should be performed if we wish to retain it. Since
this field is only used as a boolean, instead of fixing the ownership
problem, the field is replaced by an actual boolean instead.

Original-Fixed: 88807
Test: Pure refactor
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/604987
Original-Revision: 320296a69a366956b126d3aac0f996e821dcc36c
GitOrigin-RevId: fb1c46a6c51dd3098eabe361c99ee9f16b2af98b
Change-Id: I9daa14e81fd6518c940fdec6d27cdaea4ee6f7cb