[roll] Roll fuchsia [ffx] Remove common library in favor of single plugin
Sorry for the churn, but since FfxConfigBacked only works on top level
argh params, it makes more sense to just treat this as one plugin and
route the subcommands manually.  This prevents a strange dependency
graph as the subcommands will need access to the manifest parameter on
the parent subcommand.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605521
Original-Revision: 55e6c44854e183b84edb697f16a102e586271786
GitOrigin-RevId: 52715a7dcefa5ee4886810dc735cf34d4c4f410e
Change-Id: I9a57c3c9e4abd50f94d53170da6b60cb44086aed