[roll] Roll fuchsia [kernel][x86] Fix re-entrancy bug with swapgs and NMIs
This change fixes a bug where the kernel might use an invalid %gs.base
if an NMI interrupts another interrupt handler before that handler is
able to swapgs to the kernel's %gs.base.

Update interrupt_common to check whether the interrupt was an NMI and
if so save the current %gs.base on the stack and recover the kernel's
%gs.base (percpu pointer) using dead reckoning from the %rsp.

Add in-kernel unit test that would trigger the bug prior to this fix.

Original-Bug: 31612
Original-Bug: 86717
Original-Bug: 87852
Original-Bug: 88368
Test: k ut x64_platform_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/604194
Original-Revision: 01ef9cf10d8e2d975621fdfd9942739608728e64
GitOrigin-RevId: 81ac25a1c65ebc4ed47e53aa025a38b2cbd6acc0
Change-Id: I0d31491ea257d0f633a8461e066bcc8d24656664