[roll] Roll fuchsia [ffx] Try to get RCS proxy more than once
There is an issue with a test trying to run `ffx target show`. It
appears that the first attempt to get the RCS proxy sometimes times out.
Subsequent times succeed but because the logic is such that each service
needs to connect for the plugin to succeed this causes the whole thing
to fail. As the path to get the RCS proxy goes through a OnceCell, we
can add a retry in the macro, and if things actually do end up working
in subsequent attempts to get the proxy, then the retry will work right
away. The cost of this change is that we double the potential timeout
before failing because now instead of N*timeout where N is the number of
services we are trying to connect to, we would retry each of those one
extra time.

Original-Bug: 86208
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/607445
Original-Revision: 45dfd24880268d22f8eefb32733f4bcc4884c371
GitOrigin-RevId: bbb3f928d59426548aa828c4f7d71c4a68ed65e1
Change-Id: I269856ecf95cbf99811acf6f16c625e1a5b11208