[roll] Roll infra/recipes [checkout] Respect gitiles_commit + gerrit_change w/ module flag
This change introduces a $fuchsia/checkout property
"respect_gitiles_commit_with_gerrit-change", which, if set, allows for a
gitiles_commit to be combined with a gerrit_change.

When the flag is set and both a gitiles_commit and gerrit_change are
present in the build's input, the checkout strategy is to patch (with
rebase) the gerrit change on top of the gitiles commit.

Without this flag, the checkout behavior is unchanged in all cases (see
no changes in existing expectation files). This flag is never intended
to be used by default builds i.e. try or ci. See
go/fuchsia-ad-hoc-perf-runs for more context.

Note that this change alone does not satisfy the doc's requirements for
checkout. We will add another flag that changes the patch behavior from
rebase to cherry-pick.

Original-Bug: 87199
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/603601
Original-Revision: 9a124a601a5af188430c1bddea4bd64f6edbb454
GitOrigin-RevId: 697047e1701fc56ed365af631c3f17b0209f9f04
Change-Id: I9f900166052d343e6594161c0ab21e9f4666c272