[roll] Roll fuchsia [llcpp] Remove arena parameter for table fields <= 4 bytes in size
This CL removes the table field setter for fields <= 4 bytes in size
that takes an arena as a parameter. This field was deprecated as part of
the move to wire format v2 and the shift to inline field setters where
possible.

Original-Bug: 79578
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601782
Original-Revision: 2a429f6b3d55d020dc1de293bc2336bc08bb73df
GitOrigin-RevId: 081215a39a43052b8eb4a41e9933a5fbebd61f70
Change-Id: Ide55e94f1452836af4509c0ac6d45c9ccb17b135