[roll] Roll fuchsia [fidl][dart] Validate padding, disallow 0 ordinal
In the generated struct decode function, we now generate check padding
statements for all byte intervals which must be zero. In addition, we
check the padding which must be added at the end of inline structures
which are not FIDL aligned, i.e. whose size is not exactly a multiple of
eight.

We also disallow the zero ordinal in unions.

Original-Bug: 7805
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/607824
Original-Revision: 4a505c164b89915fb6e29403f5b15810c2e48a29
GitOrigin-RevId: 3d7ff3df9a161e3f26db5144d7f0f9bb53d3b68c
Change-Id: Idd1b820b4203dc04ef84778e0f422f734fe6bdc0