[roll] Roll fuchsia [rust-3p-crates] Bump rustls to 0.18.1
Updating rustls is a little challenging, since the latest version of
hyper-rustls and tokio-rustls depend on the latest versions of hyper and
tokio, which are quite large.

Fortunately we can reduce our load, because all the rustls dependencies
don't actually depend on any rustls breaking API changes. So this patch
takes rustls 0.18.1, updates it to incorporates the fix for
https://fxrev.dev/479224, and resets the version back to 0.17.0.

We also get to remove base64 0.11.0, but need to import base64 0.12.3.

Original-Bug: 87798
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601461
Original-Revision: c90d126bf2a271c01c50a000cf490e3a7509364a
GitOrigin-RevId: 53651b353817b31179ff9b08fbadcedc2c9e069a
Change-Id: I0cebda3eb896e70d0a709deceea12518c00d71f2