[roll] Roll fuchsia [licenses] Skip directories at a lower level.
We shouldn't skip prebuilt/ and vendor/ directories entirely.
This CL aims to re-add as much of those sub-directories as possible.

In order to add more, we will need to be able to skip dirs with comments
in config files, so as to not leak details in fuchsia.git.
I'm working on that now.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/608211
Original-Revision: f2258e251c7a2ed06ad92af7e056d77ad0a5865e
GitOrigin-RevId: 4f9c63088ce9c5fb8c875201e61261cb4a6fd4e4
Change-Id: I88bad9b121b95116157c00671fdf1e241a1f2b97