[roll] Roll fuchsia [f2fs] Add compatibility tests for extended inodes
If the filesystem is configured with extended inode features,
the allocated amount of on-disk bytes for inline dentries may decrease
and the way how a directory structures its inline dentries may differ.
According to that, this patch removes |struct InlineDentry| which is a
previously defined on-disk structure for inline dentries but rather
constructs instant pointers in data block pointers region when needed.

It also brings some minor bugfixes with extended inodes and adds
supports for configuring compatibility tests with Mkfs options.

Test: fx test f2fs-compatibility-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605881
Original-Revision: 4eeff9084864454844b4009319ba6bce41f301ac
GitOrigin-RevId: 35523b540d32efaca855c67fffb05a3b86935e73
Change-Id: I2d4875abf6ca207cba64c0ba621196512d28b0d0