[roll] Roll fuchsia [ffx] Adding serialization to emulator engines
Part of the lifecycle of an emulator engine in ffx emu is to serialize
the configuration to disk once the instance is started, so that it can
be read from disk on future invocations of show() and shutdown().

This CL implements the serialization for those engines. It's a common
trait that each engine will implement, so each engine need not
reimplement anything. We include a test which serializes and then
deserializes to verify the process is lossless.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/608481
Original-Revision: ee1ae4c69c62c99480d0e830510316cc96de69fa
GitOrigin-RevId: 90d9b238fe91e6aea6bfc64476ddb55997d504ff
Change-Id: Ie87d2e4b326f78ea6e57d51728a1189eca0a771c