[roll] Roll fuchsia [netstack3] Refactor inbound packet processing
Unify the calculation of how to process inbound IP packets into a single
function - `receive_packet_action`.

This commit resolves fxbug.dev/35127 by deciding to silently drop
packets which arrive from external hosts and have a destination address
in the loopback subnet. We already dropped such packets, but it was an
open question whether to send an ICMP response. This commit makes the
decision to not send an ICMP response. TODO comments related to this
issue have been removed. [1]

This commit resolves fxbug.dev/83183 by deciding, when acting as a node
(with forwarding disabled), to silently drop packets which arrive from
external hosts and are not addressed to this node. We previously dropped
them, but it was an open question whether to send an ICMP response. This
commit makes the decision to not send an ICMP response. [2]

[1] fxbug.dev/35127 reads:

    Per RFC 3513 Section 2.5.3, "An IPv6 packet with a destination
    address of loopback must never be sent outside of a single node and
    must never be forwarded by an IPv6 router."

    We already check that we don't deliver these locally, but we should
    also make sure not to forward them. Open question: Should we treat
    this as a "destination unreachable" condition and reply with an ICMP
    error message, or just silently drop the packet? I lean towards the
    latter since the originating host of such a packet is clearly and
    seriously violating the RFC.

[2] fxbug.dev/83183 reads in part:

    When acting as a host (with forwarding disabled), for IPv4, we
    follow RFC 1122 Section 3.2.1.3, which says that "[a] host MUST
    silently discard an incoming datagram that is not destined for the
    host."

    Currently, we also do this for IPv6, but there may be different
    behavior for IPv6.

Per a comment by ghanan@google.com:

    Should be the same behaviour for IPv6.

    https://datatracker.ietf.org/doc/html/rfc4443#section-3.1. Only a
    router sends Destination unreachable with code 0-3, 5-6.

    A destination node may send code 4 if the port is unreachable.

    Given the only Dest Unreach ICMP error that a host may send is code
    4 but it only applies to the destination node, there does not seem
    to be any ICMP error to send when a packet reaches a host different
    from the one the packet is destined to (none of the other ICMP
    errors in https://datatracker.ietf.org/doc/html/rfc4443#section-3
    makes sense to send here).

    Seems like we are doing the right thing.

Original-Fixed: 35127
Original-Fixed: 83183

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/572201
Original-Revision: 02ce17f053d9528f52ca26c4f865d782db71a6c6
GitOrigin-RevId: 38cf29e0c921a14e75193b70a1df4dc16f72f498
Change-Id: Ia6b721eb6172cb5fb85f6c236d065883fc9c00d2