[roll] Roll fuchsia [ffx][emu] Add QemuBasedEngine trait and stage files
Adding a trait, QemuBasedEngine, which collects the common
methods used by all engines that use QEMU as the emulator platform.

In this initial CL, the method to copy the guest files into an
instance directory for the emulator engine is added. The guest_config
is returned with the updated paths.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/608477
Original-Revision: 0002600b0151355c206c41318daeaace0a18d46f
GitOrigin-RevId: 904537feb831d76abf9c6be946c54365bc3e8a32
Change-Id: I58801684f869701b0ff9d0379375ae89f10cdc52