[roll] Roll fuchsia [fshost] move data, blob, and pkgfs to core config
This moves the data, blob, and pkgfs capabilities, and all capabilities
that route to those filesystems under the hood, to the core_fshost.cml
config. This configuration is not included on bootfs, causing those
directories to not get routed when they will not be available.

This also adds these routes to the verify_routes allowlist, adding a new
allowlist just for bootfs so we continue to verify these routes on
non-bootfs-only builds. The verify_routes tool is extended to take
multiple allowlists so we can variably add this one.

Original-Bug: 39588
Test: CQ
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569004
Original-Revision: 4c433499013d8a93a6178015cf46e2d802fe96ef
GitOrigin-RevId: f9fe31ae4263f0c6d196f73705bbc5cbc0979f8c
Change-Id: Id1ce129f1599c172e009bc2537a28378161e1b23