[roll] Roll fuchsia [gidl] Add gidl-format host tool
This CL adds a formatter for .gidl files called gidl-format, integrates
it with fx format-code, and formats all existing .gidl files with it.

The formatter is a completely separate program from GIDL, but it uses
the same lexer: text/scanner from the Go standard library. Like gofmt,
it does not attempt to split or repack lines except in a few special
cases. The style is based based on the way we were already writing .gidl
files, and leans towards the gofmt style for Go literals.

Special comments `// gidl-format off` and `// gidl-format on` can be
used to disable the formatter in parts of a file.

The name gidl-format (not gidlfmt) was chosen to match fidl-format. Its
command-line interface is also modeled after fidl-format, and it
respects the rubric (https://fuchsia.dev/fuchsia-src/concepts/api/cli):

    Usage: gidl-format [-i] [<file>...]

    Format GIDL syntax in each <file>, or from standard input if no files are given.

    Options:
      -i    Formats file in place

Test: fx test gidl_format_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/607972
Original-Revision: 2b178da1aef585989c8cf5b49ba43edbd11d0fd8
GitOrigin-RevId: 1e2faee6f75e7495ee1f74be7f855fdbdf5250ff
Change-Id: I4489c219e1e1c4972271718624370dc69bccb18a