[roll] Roll fuchsia [plasa] Hook up the plasa tests into the SDK tests
I forgot to add this stanza earlier, leading to the plasa unit tests
not running on the builders. This allowed a broken test to slip in.
This change fixes the issue.

Also needed to fix up paths to the test files for the build to become
green on core.x64-release and other core.* builders.

Original-Bug: 88914
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/608472
Original-Revision: afe268fcd5666700371fc07ae2393b92ae2980cc
GitOrigin-RevId: fb9a1ba265f8affd486438d836386eb85c1bc48c
Change-Id: I4f9e82a63e070ba2e284a7f4bb07dfb407b52a11