[roll] Roll third_party/go [fidl][go] Check for too many handles / bytes
This CL adds a new unmarshalling function Unmarshal which checks
that all input bytes and handles are consumed.

A multi-step migration is needed to migrate all repos to this new
unmarshaling function (to be done immediately afterwards).
As part of this, the existing UnmarshalWithContext function will
be renamed to UnmarshalBuffer

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/go/+/608143
Original-Revision: 1010e58fca596bcda08dedff3d94cfabc29a856d
GitOrigin-RevId: c149576c777c953327518d2b04cfadf8d1f1ff89
Change-Id: I041e93147ebbb06d6be15c4ca29362f13f9be87e