[roll] Roll fuchsia [netstack] Don't change original PacketBuffer ptr
...when cloning a PacketBuffer to write to a bridge's endpoint.

Currently, the original pointer to a PacketBuffer is erroneously updated
when the PacketBuffer is cloned. If a LinkEndpoint downstream from the
bridge makes a change to the PacketBuffer when writing the packet, all
LinkEndpoints that the bridge writes the same (cloned) packet through
will see the change. This is also a data race opportunity if any writes
occur asynchronously with the write call made by the bridge. More
concretely, if a packet is sent through a bridge with five member
LinkEndpoints where the second LinkEndpoint performs change A and a
fourth LinkEndpoint performs change B, the written buffers through each
of the LinkEndpoints will look like:

   LinkEndpoint |  Packet P's delta
  --------------+-------------------
         1      |         -
         2      |         A
         3      |         A
         4      |       A + B
         5      |       A + B

As of this change PacketBuffers are unmodified in the write path
downstream of the bridge so no issues have been seen due to this
bug. However, a later change will introduce reference counting for
PacketBuffers which means a PacketBuffer will be written to.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/609052
Original-Revision: 7210b7b6bb9a32fa8ee79e94884e01ccf6695f70
GitOrigin-RevId: 01efb91a06c90dc39f4a21d26c5866cfe365452f
Change-Id: I6b1fde2384b5e3e1c99f7fe3dd5b9fff6c6b1c60