[roll] Roll fuchsia [password auth]Move minfs mock into disk_management.
The new account module depends on Minfs so its unit tests are also going to
need to use the MockMinfs that previously lived inside
account_manager::tests. The most logical home to me seems to be along
the trait but happy to talk about other options like a mocks module.

In the future we might want to move more of the account manager mocks
too but the fact that tests depend heavily on their private fields makes
that inconvenient at the moment.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/609083
Original-Revision: 37b55997727e01455c1f24039444ae8b9144d817
GitOrigin-RevId: dc1914e3c57eebfb8e2b0ba7d6204ee0d6b39cad
Change-Id: I95a9e4aeb521fb0731b2d7f124bf1591257f63e7