[roll] Roll fuchsia [build][testing] Add more missing package deps
This CL adds a large number of the remaining package dependencies in
existing tests. It was generated by running the script in
https://fxrev.dev/606442 iteratively until no further missing deps were
found. The list was then curated slightly to account for some out of
order logs in the syslog, and updated with deps that did not match the
usual resolution failure signature.

Note that this CL still does not bring us to 100% fully specified
package deps. There are two remaining root causes that have yet to be
solved:
1. Tests utilizing prebuilt packages shipped by chrome do not correctly
specify deps, as that will require some refactoring of the visibility
rules in the associated package templates.
2. Host tests do not correctly specify package deps as those templates
have not been updated to handle this field.
Both of these root causes will be addressed in upcoming CLs, as this one
is already quite large.

Original-Bug: b/195343539
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/607345
Original-Revision: 937e26a9958fdf3aaca96ffe3043b459aa887e3e
GitOrigin-RevId: be747e469861c2b7f97373c8574aa1cbb8ee7361
Change-Id: I45175f9b778f118d298e882b8398fe1230fba50d