[roll] Roll fuchsia [audio_core][test] Fix HermeticFidelityTest for 44.1
Fix the incorrect calculation of source buffers, for SRC ratios
that require a fractional buffer length in order to translate
perfectly to the dest analysis buffer length of 65536 frames.
This is most clearly seen with test cases that include the
44.1k => 48k rate conversion, where SiNAD results would vary
by up to 12dB across runs.

Also, enable HermeticFidelityTest cases to specify that only a
single frequency should be tested. This will be most useful for
test cases that are generally frequency-independent, such as
noise-floor for uint8, int16 and int24-in-32 input formats.

Both updates are used in upcoming CL that adds the test suite
audio-core-fidelity-test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/590488
Original-Revision: 29062a46f09d237746ef85a861772f4dc2eeb4b1
GitOrigin-RevId: db4d9653e54063c7cdc783ef229cfdb9d314d201
Change-Id: I5abb0805a610d03ba44ff02c5772fb91fb1ad706