[roll] Roll infra/recipes [create_branch] Handle duplicate project names in branch initialization
It is possible for a Jiri checkout to contain more than one project with
the same name. In this case, initialize a branch for every one of those
projects, not just the first one that happens to be returned by `jiri
project`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/609711
Original-Revision: 6d553b4c1f56164948413e491aba580d7d280bea
GitOrigin-RevId: 28760ea2557c4dc2d73b47ad87511b2ff856ce76
Change-Id: I02a22f816ab4749457c5f6da1fc09766333935b7