[roll] Roll fuchsia [driver2] Forward directory node in DevfsExporter
Update the DevfsExporter to pass a connection directly to the directory
node. This allows us to then Open() subsequent connections within the
directory, and forward the connections onto driver components.

This is the recommend way to restructure this, based on conversations
with the Local Storage team.

Original-Bug: 86664
Original-Bug: 33183
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/609103
Original-Revision: 352cb12a53688d6d12b64f5bc531ca15f0ada68c
GitOrigin-RevId: 31570eb2654b8d7b69598fc8c55951659e29eb88
Change-Id: I4e25ed63cfc6998a9c746950bce15e42417a858a