[roll] Roll fuchsia [cm][storage] Wait for done signal in integ tests
The done signal indicates when our mock component has
completed writing to the provided directory. While waiting for
the signal is not necessary for the assertions made in the
test cases, the sender end of the signal in the mock panics if
the recipient end has been closed.

Original-Fixed: 89079
Multiply: storage_admin_integration_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/609006
Original-Revision: 07a446fee08a04644017b2e2d8fc53072bdccf71
GitOrigin-RevId: c299ce66e59a1308b1bbb55f869b10cb7876bc79
Change-Id: I1281742a797e69f2b8f18ee82089fae2ea416134