[roll] Roll fuchsia [debugger] Check stack for stopped thread.
Previously there were two functions to check for a stopped thread, one
that required a frame and one that didn't. However, all callers required
the thread be stopped with a frame so users of the former were
incorrect.

This converts all callers to use the "WithFrame" variant, and enhances
the checking for "stopped in exception."

Original-Bug: 89073
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/609465
Original-Revision: eebda050e408c9e1ebbe54cfc5355b3b8475c347
GitOrigin-RevId: 5ef2fc02420744f3685af5d41c0d8e488694fda3
Change-Id: I217822f8d2f5a4ff6e8e16b19e12260fd4fad165