[roll] Roll fuchsia [assembly] Add destination field to assembly inputs
Some of the assembly inputs come from the root fuchsia directory, and
are therefore prefixed with ../../. In order to consolidate all the
artifacts under a single directory, we are adding a destination field
that strips the prefix.

The Product Config will then need to be updated on the consumer-side of
this file to also strip the prefix.

Original-Bug: 87926
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/609467
Original-Revision: eeedc960442c7de19fd4f8947c2660472f691d5d
GitOrigin-RevId: f2d643fab7d8105dc3ef33e98937c278de6fba01
Change-Id: Ieac847078f28c0a8db98e65304c6ef11f3f81c96