[roll] Roll fuchsia [usb][xhci] fix locking
Originally we were obtaining the lock in a function, release it, and
then call a sub-function and obtain the lock again. This allows for a
potential race where another thread could acquire the lock before we
re-acquire it in the sub-function.

This CL requires the sub-function to be called when holding the lock,
eliminating the possibility of race conditions.

Test: tested on NUC

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/608508
Original-Revision: 6596ba0f2708c8a42f6a1d249bd3eb23198f2245
GitOrigin-RevId: d1cc96a9ac03b4a6c1b92598133ed1923d66a2a8
Change-Id: I60cb4cbee3162b4e2da2d3cd3d7d76581415b067