[roll] Roll fuchsia [network][explicit] Initial commit
Add `ResultExt` trait with `ok_checked` method. `ok_checked` is like
`Result::ok`, except that it takes a type parameter which must be equal
to the `Result`'s error type in order for a method call to compile. This
allows code to convert a `Result` to an `Option` in an ergonomic way
while complying with the Netstack team's policy on discarding type
information.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/608429
Original-Revision: 0571158fa33c277cc186aef25190650a8d8a1088
GitOrigin-RevId: 997e3d628744a5046a325f84be00be868c82348d
Change-Id: I05a97f4986bab8a1976df3b9cd6fda20a170f8a8