[roll] Roll fuchsia [acpica] Only support S5 state from AcpiOsEnterSleep.
AcpiOsEnterSleep is used by ACPICA code to validate whether we want to
transition to a particular sleep state. Since the ACPICA functions used
to transition S-states require interrupts to be disabled and in Fuchsia
ACPICA runs in usermode and we don't expose a mechanism for it to
disable interrupts, we do not want to allow S-state transitions except
shutdown (sleep state S5) where cache corruption does not matter.

When we do support other S-state transitions we will do so by jumping to
the kernel to disable interrupts and make the transition so this
function will not be called.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605221
Original-Revision: 7ac60b28c214f45bee71498f596f461f841a48e7
GitOrigin-RevId: 15d4d24c4ea5e2561fa5abd15181be8c8d3a5b96
Change-Id: Ibf41b6130e2ebb55213c3825f6e30efbd6495cb0