[roll] Roll fuchsia [ffx] Add deserialization to ffx emu engines
My previous commit added serialization to emulator engines, such that
the engine configuration is written to disk when startup succeeds. While
we confirmed the serialization is correct, it needs to be read back from
disk to be useful.

This CL adds the deserialization step, which is used by shutdown and
will be used by show (and maybe list). Some adjustments to serialization
were needed to enable deserializing before we know the type of engine,
and some error handling was added to both sides to make the interface
cleaner.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/609322
Original-Revision: 5f1621334759e51f36648ce9ed40aa53affbeb2e
GitOrigin-RevId: f997e4dfe9b0c51fcf56c3e3a41259a4946abd5b
Change-Id: I0416f83bd79f541d9952a7d93401fbadcfa004c8