[roll] Roll fuchsia [cros-ec][lpc] Add ChromeOS EC LPC interface driver.
This is phase 1 of moving the ChromeOS EC drivers out of the x86 board
driver. This CL implements the transport protocol for EC drivers. A
future CL will add a driver that binds to this driver and communicates
with the EC to provide functionality to the rest of the system.

Original-Bug: 78349
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/608118
Original-Revision: 85b46e667cce4d06ed0d0d17f3fe86c2773b652c
GitOrigin-RevId: 51d5a2b6a23e9670d58f97615fdddbf3cd5f1f3d
Change-Id: Ie71015855289f55fc727ad075b82d6c54be7d181