[roll] Roll fuchsia [audio][test] Adjust fidelity frequencies
In our fidelity testing, we use sinusoid input signals to characterize
a pipeline's frequency response and signal/noise. We measure frequency
response and signal/noise at specific frequencies, enforcing that the
results meet our expectations. Our set of test frequencies span from
subsonic (including 0hz) through the audible spectrum into the
ultrasonic range.

We get more accurate measurements because we avoid "windowing" the
output. This is possible because we use specific test frequencies that
lead to an integral number of output wavelengths in our power-of-two
buffer. We have discovered a side-effect of this test optimization.

Test frequencies that share many common factors with 48000 are prone
to slightly inaccurate results (+/- a few dB) compared to the rest of
the spectrum. Changing the frequency by even 1hz in either direction
leads to results that are much more consistent.

This CL changes our test frequencies to avoid frequencies that are
multiples of 375 (the non-power-of-two factor of 48000), specifically
21khz and 31.5 khz.

We still believe this no-windowing optimization is the best approach,
this slight workaround notwithstanding.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/609979
Original-Revision: 8208726268e1755a89f031237454fd743dd5e622
GitOrigin-RevId: 91bccca0ad1093d13882321758b66fd175e4ac0b
Change-Id: Ie78a8b0d5adf74234b4eca11fb41840af735c185