[roll] Roll fuchsia [fdio] Ignore MSG_NOSIGNAL
Commit 049ac5352f60b670f3bcac046731fc85eedbd290 defined this flag but
neglected to account for the fact that its definition would cause
downstream consumers to pass it to functions on the send path. Since
signals aren't implemented in Fuchsia, we can safely discard this flag
if it is passed.

Update various tests to pass this flag when it is defined, rather than
when __Fuchsia__ is not defined.

Original-Fixed: b/207057734
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610383
Original-Revision: 1a46834d23750e4f5db7be29e8b3b0bd2162aeb0
GitOrigin-RevId: e2c41b28e745c70943caad85f3123e4a073cd105
Change-Id: I3e4d2fb2333ea1ccb18333886f724eb99d555860