[roll] Roll fuchsia [ffx][driver][lsblk] Add lsblk subcommand
The fuchsia-vfs-watcher library was also updated to work host-side. A
macro which in fuchsia-zircon-status was made public in order to make
this work, and a struct that was defined in fdio_sys was copied to avoid
dependencies that only work on fuchsia.

This doesn't implement support for lsblk subcommands - those should be
implemented as separate commands.

Tested: ffx driver lsblk
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610022
Original-Revision: f4e5b209f4eefe1e3f0c9371b9f2dbcdff747bf7
GitOrigin-RevId: 954398542a88ca3d9ff16cfd4f0cc8cf9a36fdb9
Change-Id: Id6201a75196c17c1c627b45be8acdfd0e30c88bb