[roll] Roll fuchsia [ffx][repo] Use SSH_CONNECTION for non-tunneled connections
Ideally we'd prefer connections from the target device to the ffx
repository to go through the ffx tunnel, since that would transparently
work when tunneling through a local device to a remote workstation.
Unfortunately the native ffx tunnel has a significant bug
http://fxbug.dev/87439, which can cause resolutions to fail.

To avoid that bug, this patch changes the repository configuration to
only create a tunnel if the server is running on a loopback address.
Otherwise, we'll instead tell the device to talk to the host address as
seen by the environment variable $SSH_CLIENT. This matches how
scripts like `fx serve-updates` configure devices to talk to the pm
server.

The downside of this approach is that it does not transparently work
well with scripts like `fx serve-remote`. Those scripts will need to be
updated in order to forward the repository address at the ssh level in
order for it to work.

Original-Bug: 87439
Original-Fixed: 88761

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605986
Original-Revision: 70ae151518e131780a2774cc33c62a84612f0923
GitOrigin-RevId: 8d32a737e8ea7bb303d35a4d0173b4649f7d0a6f
Change-Id: I07d4451bcbff683e17a08b7b2ccdaffff183333c