[roll] Roll fuchsia [ffx] Implement get_all_instances and ffx emu list
This CL adds the implementation for the get_all_instances function in
the ffx emu engines module. This function is currently used in the
"list" subcommand, to provide a listing of all of the current instances;
and also in the "shutdown" subcommand when the user passes in the
"--all" flag, to clean up all instances on the system.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610369
Original-Revision: 3e221ebaef94eb1da6c708fc87c7a49237ca29ad
GitOrigin-RevId: b4a78242d429c5853db3a70dd42c93faa2507f3c
Change-Id: I31b2ba8ab6d8fb029dc4b2764ac4de3757caf79e