[roll] Roll fuchsia [bt][hfp] Give `a2dp.Controller` to integration test
Context: The `a2dp.Controller` capability is used by the HFP component
to pause A2DP media when needed. The HFP integration tests do not rely
on this capability yet. However, the HFP component opportunistically
connects to this capability. Consequently, when running the tests, a
warning occurs because the `a2dp.Controller` capability is not
available.

Give the integration tests the `Controller` capability which is provided
by an existing mock.

Test: fx test bt-hfp-audio-gateway-integration-tests, verify tests pass
+ no warning

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610128
Original-Revision: 49bd9e0a1763e10defeca987388b4227ee4a4f12
GitOrigin-RevId: a47741f25c0207b984a343357b42292e4b5fe87b
Change-Id: I9d4afd0bbf04459edefffc1e4659b9120b10a096