[roll] Roll fuchsia [driver_manager] Append driver-related items to mexec data ZBI
There are ZBI items that the driver framework expects to be present in
the data ZBI that are not inspected by the kernel. Accordingly, now as
the canonical mexec caller, it would be more reasonable for
driver_manager to be responsible for appending these items before
mexec'ing (instead of this happening within the context of
`zx_system_mexec_get_payload()`).

Compatibility-wise, it should be safe to make the userspace and kernel
changes at the same time: an older version of userspace is not
guaranteed to work with a newer kernel - and, at worst, a newer
userspace that performs an mexec with an older kernel will have
duplicates of the driver metadata present in the ZBI.

Original-Fixed: 89185
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610064
Original-Revision: 95b6125184286e5bddf5372760b4a3bfb34d9cc7
GitOrigin-RevId: a0babfbfbfecac1cdbdafcc401e1e98d70744fc2
Change-Id: I4f6bfdafa425daba5d1647b179dbeded24482983