[roll] Roll fuchsia [ffx] Implement a basic Show subcommand for emu
The last piece of the ffx emu interface is Show. It displays the details
for a given emulator instance, such as its status and the configuration
used to create it.

This CL implements a rudimentary Show subcommand for the Femu and Qemu
engine types. This subcommand just dumps the configuration to the
terminal using a debug string. It's not formatted in any particular way,
just uses Rust's default Pretty Print, but it serves as a starting point
for understanding what we still need to provide to make it a useful
tool.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610542
Original-Revision: cc1d6afc47f5454921c51f7e6fc183976021a91c
GitOrigin-RevId: d50e89f79b3c97e945008179113eb3485aae4e56
Change-Id: I5b9d297c6d874004c124f05fcb9c3ac5c6737694