[roll] Roll fuchsia [core][test] Avoid mappings being merged
There is a chance that when randomly placing mappings in a VMAR the
kernel happens to place a mapping in such a way that it is a candidate
for merging. This has been possible for a long time, but required the
two mappings have the same permission. Now they can have different
permissions and be merged, greatly increasing the odds of this flaking.

The fix in this CL is to map in every second page of the VMO, ensuring
that mappings can never be merged as they are not contiguous in VMO
object space.

Original-Fixed: 89314
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610568
Original-Revision: 7e0ca9d39bea2ca275f26dd4432785951a400361
GitOrigin-RevId: 5df796a11f718aa9090d771d06edf6eaf69de2df
Change-Id: I4c66978a0d49ff27b30753ff44d54ffb67be9253