[roll] Roll fuchsia [fxfs] Fix assertion that could fire during shutdown
The corrupt tests would occasionally show evidence of a panic (without
actually failing), due to reservation amounts getting out-of-sync when
futures got dropped at shutdown.

This CL fixes it so that futures should no longer get dropped in the
middle of a transaction.  This involved fixing VFS so that connections
are only terminated between processing requests, and fixing an issue
with the journal flush task whereby it would drop the compaction task if
it encountered a flush error (which the corruption tests would trigger).

Whilst debugging this, I took the opportunity to refactor Reservation so
that it supports sub-reservations (rather than holds) and supports
reservations where the owner is something other than Arc<dyn Allocator>
(it's now possible for it to be a reference).  I renamed the methods
that require some care to "forget" to match Rust's function of a similar
name.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/609107
Original-Revision: c10989419adffdf3e81019c1b46e616e36770e71
GitOrigin-RevId: 67c3ee6af83c096626feeec4b36c7b9bf72dfb01
Change-Id: Ia7c550e63eaba224cbd524aa975e7c93b29495df