[roll] Roll fuchsia [fidlc] Overhauling constant resolution
We can start with some eye candy:

    error: "abcd" (type string:4) cannot be converted to type string:3
    const s string:3 = "abcd";
                       ^~~~~~

    error: 256 overflows type uint8
    const u8 uint8 = 256;
                     ^~~

    error: 42 (type untyped numeric) cannot be converted to type bool
    const b bool = 42;
                   ^~

This change overhauls how constant resolution is done to leverage type
inference of literals, and check for convertibility against the target
type sought while resolving.

Traditionally, it is preferred to have the inferred type bubble up from
the leaves up to the roots, while at the same time having a "type
constraint" flow from the root to the leaves. This makes it possible to
check the inferred type against the constraint when appropriate.

The approach taken here is to combine constant resolution along with
type inference. We modify `ResolveConstant` to optionally take a target
type, i.e. if the type is provided we resolve and type check, if the type
is omitted we resolve and infer the type.

When inferring the type of numerics (say `3`), we can try to infer the
smallest type(s) (here `int8 | uint8`) and then have convertibility rules.
That tends to be complicated, and also pushes the language towards
implicit conversion which we would like to avoid.

Instead, we infer numerics to a special `untyped numeric` type which is
convertible with all actual concrete numeric types (like `int8`). This
makes the type system a little more flexible, though it requires
checking for overflows in addition to type convertibility when dealing
with numerics.

Appendix
--------

Some history about type inference in the compiler, its introduction,
deletion, re-discovery with a problematic implementation, and its
re-introduction here (hopefully to stand the test of time):

* We first added type inference in
  I70dd8eb79dd2b535b9e8f44330c18f9bc427c001. At the time, the compiler
  did not properly check `const` declarations, and it was natural to
  rely on type inference for type checking, i.e. infer the type of a
  literal, ensure it is convertible to the type it is assigned to.
  This change also clarified and ensured that enum values be treated
  nominally, i.e. `MyEnum(3)` is not convertible to `uint32(3)`.

* However, there were some out-of-order constant resolution issues.
  These were addressed in Ib6491a14c677d03bea618b19f68a88993588efde
  which approached `const` declaration checking using a more brute force
  style. This worked well, since there was no need for type inference
  per se, rather, type inference was used as a tool to build `const`
  declaration checking.

  Fun fact: It took me a `git bisect` to locate this change. I
  remembered that type inference had been added, but had forgotten why,
  and when.
  Fun fact #2: Both changes introducing and removing type inference
  landed after being first reverted. Will this change suffer the same
  fate?

* With the introduction of fancier attributes (see RFC-0086), the need
  to infer types became a reality. Say you write
  `@my_attribute("hello")`, the compiler must infer that the constant
  provided is a `string:5` and confirm that only `string` or `bool`
  constants are ever used. This was addressed in
  I890573b74c582b7c3c74d4ac602ed6b6b976c09b using an error reporting
  silencing technique: try to resolve constants with errors off, it
  works? that type was good, it failed? try with another type. As we
  quickly saw, silencing error reporting is too coarse of a tool in a
  compiler, and leads to papering over issues
  (Ie0c7e0b3297f91fc8967baf7dc3c59b3d1b6f227 documents those).

Which leads us to this change! The return of type inference, hooray!

Original-Fixed: 87521 (crash on cycles among attributes/consts)
Original-Fixed: 87614 (wrong error for unknown identifier in attribute arg)
Original-Bug: 89330 (precise inference of string type bounds)
Original-Bug: 76219 (type canonicalization)
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610441
Original-Revision: 9cb88f2e7f874703e0ead9ccdc2c0cdc446dadd4
GitOrigin-RevId: ade7c9f80a4f96c162ca1cf342f0ee025b707c5c
Change-Id: I8ea50d7003ac54d6697a596cf90d3dfe2214cf00