[roll] Roll fuchsia [realm_builder] compare resolver decls against pkgfs decls
Some builds automatically insert a use declaration for the DebugData
protocol into component manifests. This is an issue in tests that are
expecting to know the contents of the manifests on disk. Work around
this by having the tests compare the contents of the realm builder
resolver against the manifests as they exist in pkgfs, instead of
comparing then against hardcoded values.

Original-Fixed: 89446
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/609862
Original-Revision: b24aab37002b680e9f7bb3adc9f05d526597f44e
GitOrigin-RevId: e4a614dc6414ac1367eab2395fe38ffc470e148f
Change-Id: I280d6d1537c1876ba1806d7ea173fd46473e269c