[roll] Roll fuchsia [settings] match watch to watch2
As part of the migration process for settings.Input, the clients
currently on Watch2 and SetStates will be migrated to the more
properly named Watch and Set. For this to happen, the underlying
functionality of Watch and Set will need to match the current Watch2 and
SetStates. Watch2 and SetStates will be removed after clients are
migrated.

Original-Bug: 65686
Test: fx test setui_service_tests, fx test setui_client_interface_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/603143
Original-Revision: 454e1afa3103d794053c20f9b6850e2a518b997d
GitOrigin-RevId: f9ad146d7fb36a058e9580639f8e16d2910ce14c
Change-Id: I5ff7d3819be774500a31daddaaecc989800bd681