[roll] Roll fuchsia [ffx][config] Don't save if config did not change
There is no locking between the ffx daemon and the ffx CLI, and so they
can race each other when saving to the config. To avoid that, this patch
changes ffx config when setting a value to only save the config if we
detect that the value has actually changed. This will reduce the writes
to the config, and lower the chance that a race will occur.

Original-Bug: 81502
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610773
Original-Revision: ef16590205631300ec4ab2b5c5bf1f6b81e1043e
GitOrigin-RevId: eac186bbe2e901b7cb8f6ee2454444110fccb0d9
Change-Id: I24b3ca909aae99fafe87361ba96215fb0959ffa8