[roll] Roll fuchsia [subprocess] Avoid copying lock
Pass a pointer to the mutex instead of passing it by value so that
`handleSubprocessCleanup()` correctly shares the same mutex as
`RunWithStdin()`.

This was caught by `go vet` and should have resulted in a Tricium
comment, but `go vet` in Tricium is broken (fxbug.dev/89469).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/611721
Original-Revision: d8e797713fef1d749a888b86b421d1e1684e9017
GitOrigin-RevId: 755d21418436d982e47b4a2a8099f4136b83bd20
Change-Id: If6cb401aeabe4eb7d8ef93fcbf7e9d73ac322438