[roll] Roll fuchsia [overnet] Wake quiche writers that return Err(Done)
When quiche returns Err(Done) from stream_send(), it may not have
actually established the stream. This can mean that when we call
writable() on the connection, we won't ever get that stream ID.

Since we go to sleep when we get Err(Done), and we use writable() to
determine when we wake up again, this means writing stalls forever.

To fix this, we add a new set of wakers that is always woken when the
connection becomes writable without checking if a particular stream is
writable. Err(Done) adds us to that waker set rather than the usual
waker set for writers.

This may fix a few connection stall issues we've seen previously, and is
absolutely necessary for the Quiche 0.10 upgrade.

Original-Bug: 79699
Original-Bug: 87439
Original-Bug: 87372
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/611746
Original-Revision: cf8dfcd391210bb71b919280f2c9f4e99090eba1
GitOrigin-RevId: da46811a3a6aa6f786b6684eb54b20c8b60966b1
Change-Id: Id561fa7bdf7078f2eac28d4aebefdd391c7f6c1d