[roll] Roll infra/recipes [fuchsia_cipd_roller] Use checkout recipe module
For some reason the Jiri checkout in the run_script.py recipe is able to
check out the smart-integration repo, but the raw Jiri commands in this
recipe are not able to (see ci.chromium.org/b/8829194485082630833). I
have no idea if using the same `api.checkout.with_options()` function
will fix that, but it seems worthwhile anyway to simplify this recipe.

Expectation file diffs are a result of:
- Adding a new step that ensures the checkout dir exists.
- Resolving the manifest repository revision ahead of time instead of
  letting `jiri import` resolve it.
- Several inconsequential flag flips.

Original-Bug: 88987
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/611624
Original-Revision: 576ae5062402bb865da0744985d9642f1f93b3d8
GitOrigin-RevId: 248080675adf669193c0f8dd07039699a5d3ca9a
Change-Id: I360420b2fba455c4318bb1137a843e22943718cb