[roll] Roll fuchsia [ui] Move terminal rendering logic to a library
This library will be used to share code between virtcon
and the Terminal app.

This also fixes some cursor rendering problems that were
a result of assuming that the cursor and cell content
would not occupy the same cell.

Includes a new set of a unit tests this rendering code.
This was made possible by making some small adjustments
to carnelian to reduce the complexity of creating a Mold
context for testing.

Constants were also added to carnelian that makes it
easier to determine how to limit the size of the terminal
in order to not exceed the current layer limit.

Original-Bug: 74628
Test: terminal_lib_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610386
Original-Revision: c7600fa2cb4fca7bb07c12117bc50307aa0ca93b
GitOrigin-RevId: a09233662a7cbd1fcd01e2e90006ee6f2a8ae6b3
Change-Id: I026c541e2287967bd568fc082c4cc1ee3472be14