[roll] Roll fuchsia [build] Check for Go files missing from `sources`
This would have prevented the problems that I fixed in fxrev.dev/604261
and fxrev.dev/603765 where we had forgotten to hook some Go files into
the build graph and some tests had bitrotted.

We also now enforce that all `sources` are in the root of `source_dir`
rather than in a subdirectory. There's no reason to specify `sources`
from a subdirectory of the BUILD.gn file's directory because
`source_dir` can be set to the subdirectory instead. This removes an
unnecessary degree of freedom from setting up a go_library, and also
makes the gen_library_metadata logic a bit simpler because it can assume
`sources` are all in the root of `source_dir`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/606127
Original-Revision: 5b7b73a36f860e1377078c1bcbe6d7e74320820b
GitOrigin-RevId: c39a8dc3f3ecfa8043ee90873a0d2b1363ba70fa
Change-Id: Ib7de2c422bca93399989b80891b5492412113392