[roll] Roll fuchsia [memalloc] Introduce Pool::size()
This change introduces an O(1) size accessor for memalloc::Pool,
returning the number of tracked, normalized ranges. This will be useful
for when we get into the business of allocating buffers whose size will
be bounded by or be a function of the size of these ranges (e.g., with
physboot passing coalesced ranges in its hand-off).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/611723
Original-Revision: b5268b8b07887c185834d77206e1e2c46545696f
GitOrigin-RevId: 2a6d01569b27ab57c96bfd077652adad29b51e91
Change-Id: I7e376d6250f6f5739fc1d8fc7429d231fdc03a7c