[roll] Roll fuchsia [x86][acpi][ec] Only acquire global lock if it's necessary.
Previously, the EC driver contained an assumption that _all_ requests to
the EC required the global lock. This is only the case if the "_GLK"
method exists on the EC device and it evaluates to 1 (ACPI v6.4, section
12.8 says "An embedded controller will require the inclusion of the GLK
method in its ACPI namespace if potentially contentious accesses to
device resources are performed by non-OS code.").

This CL modifies the EC driver to only acquire the global lock if the
_GLK method exists and evaluates to 1.

Original-Bug: 87281
Test: boot atlas and check EC-related functionality still works.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/611882
Original-Revision: 426182d6de210e31948c7f0cf2d16bbd3a17be8f
GitOrigin-RevId: c7fe65829ee5aeb0a25d450dfc7bfb9db8e2f734
Change-Id: I7e9dbe0619ed78a3393e49cc7de03ad9f80027b4