[roll] Roll infra/recipes [recipes] Emit cached revision as an output property
Our incremental build statistics currently use the presence of an
incremental cache on a bot as an indicator that a cache was found.
Unfortunately, this is not always true, as sometimes the cache can be
empty. For example, if the previous build failed, the cache will be
purged and left empty. This CL modifies the checkout module to output
the cached revision as an output property, which will allow us to more
accurately query whether or not a build found a full cache. It may also
allow us to do more interesting things like measure the distance from
HEAD.

Original-Bug: 87277
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/612189
Original-Revision: 91e75749464d20e5e032506d27908a67224c2acf
GitOrigin-RevId: d9bb15f9edb287151d0e963075498f64b5bdc59a
Change-Id: I5dbfa605b5e11049536fa86f00444a7c0b5a3e5e