[roll] Roll fuchsia [realm_builder][rust] rename ChildProperties to ChildOptions
The ChildProperties struct was introduced to bring the Rust client
library closer to the API outlined in RFC-0140, but this landed before
the RFC did and the name was then changed late in the RFC process.

Now that RFC-0140 has landed, update the name in the Realm Builder Rust
client library to match the finalized API names.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610282
Original-Revision: 98ce7b9de404eb22bb97f27909d61be0418e9c40
GitOrigin-RevId: 94f964f195fcad4c066ea6ae9ad4ce74900497b9
Change-Id: I4482c76d268a15a94cfd40eeea77d239ab43a3e1