[roll] Roll fuchsia [fdio] Do not allocate 65k on the stack
In commit 299662b6df7da08a178792d8951d788c23138877 fdio was migrated
from a struct-based FIDL API to a table-based one; one of the
consequences of moving to tables is that FIDL clients must always
allocate the maximum size of a channel payload before reading, which is
pathological when the allocation takes place on the stack, as it does in
fdio's getaddrinfo.

Remove the stack buffer so that allocation happens on the heap instead
of occasionally stack smashing.

Original-Bug: 89478
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/612421
Original-Revision: 2fcd26eaeaaffd2c04fe3d612cadb30066cb9d4f
GitOrigin-RevId: 97acc6ff5ca1ae4877eee360381ba512c61ba2a6
Change-Id: Icc71c8286bd9c317b3b4fa3f87fc51948ce0e0ff