[roll] Roll fuchsia [wlan] Create a components library for drivers
This change creates an initial set of basic primitives usable in wlan
drivers that implement the generic network device interface. These
components initially center around the basic transmissible, the frame,
and its storage and lifetime management.

A frame is a collection of metadata relevant to transmission and
reception of frames. Notably it does not contain, allocate or manage the
lifetime of any memory containing the frame data. Instead it contains a
VMO identifier and a pointer to where that VMO has been mapped into
memory.

In addition to a frame object there is a class that manages storage and
lifetime for frames and a class for passing around a collection of
frames.

Original-Bug: 79622
Test: Included unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/561101
Original-Revision: 5d0ccf47d88d85e12a07912b8afa43f4a8a1b88c
GitOrigin-RevId: 3be8deaf0d22b618f7598aece4d8d8a61f201ada
Change-Id: Idbad97439d48b327bf172e5b3cb74264e303fbcc