[roll] Roll fuchsia [wlan][mlme][wlanif] Scan request fixes for channel and SSID lists
The duplicate ssid field is removed in favor of a single ssids_list
field for scan requests. Also, the channel and SSID lists are no
longer arbitrarily limited in mlme and wlanif.

This change also makes it an error to submit an empty channel list in
a scan request. That behavior is not currently used since SME always
submits a non-empty list in practice, so this does not result in a
behavioral change.

Original-Bug: 88649, 88650
Test:
1. fx test wlan-sme-tests wlan-mlme-tests \
  fuchsia-pkg://fuchsia.com/wlanif_unittest#meta/device_test.cm \
  fuchsia-pkg://fuchsia.com/brcmfmac_sim_test#meta/escan_params_test_component.cm \
  fuchsia-pkg://fuchsia.com/brcmfmac_sim_test#meta/passive_scan_test_component.cm \
  fuchsia-pkg://fuchsia.com/brcmfmac_sim_test#meta/active_scan_test_component.cm
2. fx test //src/connectivity/wlan
3. Manual save and connect to network on Astro succeeds.
4. ACTS smoke tests pass
5. Manual save and connect to network on Atlas succeeds.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605366
Original-Revision: 40860987335e23d99484c692baacdc51da873951
GitOrigin-RevId: d47ee3045d4ab688a289cd4da8104793d07d441f
Change-Id: I33e6e39b02d62febc87c223d6aa7d9b803e2f622