[roll] Roll fuchsia [fidl] Refactor JSON IR to store method payloads as types
Previously, the JSON IR stores "maybe_request\response_payload"
as a simple string pointing to the underlying struct type of the
payload. As we move to using other layouts (namely unions or
tables) as top-level method request/response payloads, this will
be insufficient for most backends, as they will need to
introspect about the underlying type of the payload, rather than
just assuming it is a Struct populated by StructMembers.

This is the first of several changes refactoring the JSON IR,
with the eventual goal being the removal of the (deprecated, but
still widely used) "maybe_request/response" fields, and the
removal of "maybe_success_variant_struct," which will be
unnecessary in a world where all payload shape information is
stored as types rather than structs.

Test: fidlde regen fidlc
Test: fidldev test all
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605061
Original-Revision: d32ac1c386eaeb347c55bef41bc8fb67ad51510f
GitOrigin-RevId: f4e0d3629fb50c26e7e221dd6d2fc67ada5b7032
Change-Id: Ifd06ed2b3938aa906b3aa16a1b5db0e17eb07fbe