[roll] Roll fuchsia [kernel][phys] Consume UART driver information from hand-off
This change hands off the UART driver information from physboot to the
kernel, in the form of uart::all::Driver. While the kernel is not yet
able to actually use the driver as handed off, it is able to make direct
use of the encoded, associated metadata, forgoing the ZBI parsing
otherwise needed to obtain this info.

Original-Bug: 88112,89182
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/609721
Original-Revision: 99ed10c264b67aa7e03a7a3fe6bc86ad08756847
GitOrigin-RevId: 84200f1feb3c233255074938c1684aa305af829d
Change-Id: I229e84874a0d5ee518167ae4bcec30b39da061a2