[roll] Roll fuchsia [fidl][hlcpp] Remove object_coding.h
These APIs are not aware of the wire format version.
The DecodeObject function incorrectly tries to parse the first
few bytes of the payload as a FIDL message header.

They are not used, so removing them is an easy way to avoid
future troubles when switching to the V2 format.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/612501
Original-Revision: 53c35301d7e46365cab5dfed52177cfa7b55fa9c
GitOrigin-RevId: b1782ccbb54bc7edf18ffe77150a951d3f73de0a
Change-Id: I3a22daece4342f4538e3fd269f6b60dd36adcc7c