[roll] Roll cobalt Change duration type in test case
When tests are compiled with the -Werror and -Wctad-maybe-unsupported
flags in internal builds, the test fails to build with the error:

"error: 'duration' may not intend to support class template argument
deduction [-Werror,-Wctad-maybe-unsupported]"

This is due to the std::chrono::duration not specifying a template type.
Instead, change the duration to the same unit returned by the Interval()
function.

Original-Bug: b/207542855
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/612524
Original-Revision: 80321e373658c02cf70727d697ecc19271d1f0ce
GitOrigin-RevId: 70d905399c5c665ff6ca4b70fa159230b50e1051
Change-Id: I03cf7cd3b2b94c568d881489e67bde55c52582d4