[roll] Roll fuchsia [llcpp] Add Transport parameter to BindServer and related functions
After this CL goes in, it should be possible to define servers for
alternative transports.

While there are no tests in this CL, building with an alternative
transport was tested with an in-progress CL introducing an integration
test.

Original-Bug: 85734
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/606041
Original-Revision: ff95bfd9a38672caecb31c2a6795a8c874da8a99
GitOrigin-RevId: a6e33c59fc20eb073e501389ab7f356c4c334ca4
Change-Id: I53383086356a4f9bcf7229afa7eb5d55b6b649c8