[roll] Roll fuchsia [build][drivers] Enforce drivers info file inclusion
This CL makes it a hard build failure for a public driver (defined
as drivers in //src and //zircon) to not provide an info file.
An info file is automatically created by fx create driver or it can
be manually added.

pshickel@ suggested we document this requirement in
fuchsia_driver_component.gni, so we add comments stating this.

Test: Removing an info file for a driver in //src results in a build assertion reporting "Must include driver info"
Original-Fixed: 81799
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/612132
Original-Revision: 2d540b4facc3a90fa3a618fb4fdd94e1ac7a3f79
GitOrigin-RevId: 1c48b170188c4ab4e1a4f975c06e7233d1017559
Change-Id: Ied9c59430b1cecac73c715b1d159afcb4cfbfb4b