[roll] Roll fuchsia [fidl][cpp] Minimal struct domain object
The intention of this CL is to support using a standalone struct as
requests/responses. Introducing a natural struct allows us to generate
request/response payloads as regular domain objects in a later CL.

The example code was updated to demonstrate some of the niceties in the
new API, such as accessors and designated initialization (a la named
arguments).

fxbug.dev/82189 is a bigger umbrella ticket to fully define the whole
set of natural domain objects.

Test: fx test //src/lib/fidl/cpp

Original-Bug: 60240
Original-Bug: 82189
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/612982
Original-Revision: c6b88adf80b23f59a1442e76f574628ed1830ffb
GitOrigin-RevId: bc49cd62557958d9388b47c5a0ff2658394605fa
Change-Id: Icaa7b0ca04b4f7afa1f61ab6c68ddb0b6384af12