[roll] Roll infra/recipes [fuchsia_cipd_roller] Add property to fail is packages are stale.
If max_stale_days is set, the roller will fail if one of the manifests
to update has nothing to roll for more than max_stale_days. That way we
can notify on failure if frequent rolls are expected but nothing has
been rolled for a while.

Original-Bug: 89709
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/613721
Original-Revision: 2bcd0400eaa074e40329ee29e8ce317297901273
GitOrigin-RevId: 4889d4174f617d027e112bdb5725a9edbfcac2ac
Change-Id: I90edb5c70b38319ee5e24fe931e03a7ec72aa3ab