[roll] Roll fuchsia [component_manager] reduce log level in storage admin
The storage admin protocol will issue a warning log if it fails to lock
the resolved state of a component. This situation can be handled
gracefully, and may be occasionally encountered during component
destruction, so the `warn` log level is too high. Reduce the log level
to `debug`, so as not to alarm users who see the log.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/613643
Original-Revision: 56c15d3b926220cf32620a9657f9748dd7065f48
GitOrigin-RevId: 35aa9fe407efe71a7b18e6f844f7d3ce7e8492a7
Change-Id: If6722f53185c21d679143f478766d6e88d00abcf