[roll] Roll fuchsia [ffx][test] Verify structured output conforms with schema in tests
It's easy to update the format without updating the schema. This adds
a check against the schema whenever the output format is deserialized
in tests. This technically violates the locality principle outlined by
the testability rubric, since we'll be validating the 'directory' crate
using tests in 'run_test_suite', but it will help us catch in the
future if a test case in 'directory' is missing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610867
Original-Revision: 397a303d0087b3bcb721258c46f316e9519b9a28
GitOrigin-RevId: 28c38ef189e8f48de82b3758dc1977da9cbb141e
Change-Id: Ida2bdc4e39e10f5f881f398278eeacf89818db9e