[roll] Roll fuchsia [input] Adds LockState to the KeyEvent API
`LockState` maintains the state of the lock keys, for cross-runner
consistency.

`Modifiers` maintain the actuation state of the keys.

Added the definitions of some common modifiers and lock states
so that we can implement a handler that maintains those.

Shortened the imports to reduce the visual clutter in the code.

The modifiers are not yet forwarded to clients.  Hooking them up is
relegated to a future commit, since this one is already getting
kind of large.

Original-Bug: 81915
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/612021
Original-Revision: 1fea70538e292c51316e5846cbb24bc8073885f6
GitOrigin-RevId: ed1da8ce8a15526e99b782367d9e56687e09a7b8
Change-Id: I97d372c5885af1b95aeb1a283ca3038373fe197f