[roll] Roll fuchsia [fdf] Add DFv1 / DFv2 interop integration test
This integration test verifies that we can load a DFv1 driver through
the compatibility driver in the middle of a chain of DFv2 drivers. This
shows that a DFv1 driver can be both a child and a parent of a DFv2
driver.

Original-Fixed: 86512
Original-Bug: 33183
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/613548
Original-Revision: 00f786715a9200dea64cb0be45858c3b65329936
GitOrigin-RevId: 5b95f465326840a78b1e0a9193db72fac356bf88
Change-Id: I56c599a658f260f08b213710edd19f181b329e5d