[roll] Roll fuchsia [starnix] Add vulkan initialization code to starnix
This code creates vulkan instances and devices, and sets constraints on
a shared buffer collection.

The code will be tested by running the unit tests for minigbm, but those
tests are not currently available in a form that can be run on starnix.
The change has been tested locally, and eventually the test binary will
be run as part of the starnix test suite.

Original-Bug: 89887

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/613302
Original-Revision: 21823e27bf2e249ac4bbea4711b72f2d08f7579c
GitOrigin-RevId: 741e4be3905f9df29ec329d77978e6cfd673f309
Change-Id: Icaa976147c6fba801341b367fee33b6e0272d99d