[roll] Roll fuchsia [kernel][vm] Package pager writeback core tests
The pager writeback core tests only run in the unified mode on bringup
builds. Package the test so that it can also run as a standalone binary
in core builds.

Ideally we would want to have the next vDSO available for standalone
tests on the bringup build, which is tracked in fxbug.dev/89597.

Original-Bug: 89597
Original-Bug: 63989
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/608543
Original-Revision: a2a25daf10480bc323c0d73e9b026007947c8b42
GitOrigin-RevId: 663f5ceda419e04c76e4b49957dd897e580f7bfc
Change-Id: Iab97d13aaa670023b53c34f72a5dc172a43fc0e6