[roll] Roll fuchsia [lowpan] Move types out of lowpan-spinel-driver to lowpan-driver-common
In the process of developing a new LoWPAN driver that directly uses
LoWPAN, it seems I need a lot of the types that were previously only
present in `lowpan-spinel-driver`. This CL moves these types to the
shared `lowpan-driver-common` crate so that they may be used by the
future `lowpan-ot-driver`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605984
Original-Revision: 85cf47932359ea76deef555e5ee0c79422c5ae98
GitOrigin-RevId: 15b438fb00e245e14df3906c733d35c0f60a8741
Change-Id: Ia9bceb10cc33151a9f6a097c6a817122c7974999