[roll] Roll fuchsia [fidl][llcpp] Allow loop shutdown to teardown a WireClient
Several folks reported a pain point when using WireClient: it's
impossible to correctly shutdown an async loop when a client is running
on it.

This is due to the thread checking performed in WireClient in
preventing use-after-frees: on most operations, we check that the
current thread is the same as the thread which created the WireClient.
This ensures that the binding never makes a to-user call while the user
object (along with the client) is destroying, since those operations are
forced onto the same thread.

In the case of async loop shutdown, the shutdown must be performed by an
additional thread, which interrupts and joins any worker thread running
the async loop before notifying the binding of shutdown. This trips the
thread checkers since the shutdown thread is not the same as the worker
thread.

From the perspective of preventing use-after-free, a single-threaded
dispatcher will not be able to run any parallel code while it is
shutting down. Therefore we can relax the thread checking rules during
shutdown to fix the aforementioned pain point.

Added a note on the LLCPP threading guide and also updated it to mention
another use-after-free risk covered by an email
(https://groups.google.com/a/fuchsia.dev/g/fidl-dev/c/peLGiIKrufQ?pli=1).

Test: fx test -vo fidl_llcpp_dispatcher_tests
Multiply: fidl_llcpp_dispatcher_tests
Test: staged the docs and verified they render correctly

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/612137
Original-Revision: c2b3976c08ea5b9e438959563f3d513942b7b224
GitOrigin-RevId: b38042868b35457718015e7f4c397eebd4910cc6
Change-Id: Iaf1b6f49d3a6210a041caca60f8085ab3fae64f5