[roll] Roll fuchsia [fidl][cpp] Minimal support for structures encoding
Adds a `foo.Internal__Encode()` method to the struct domain object
which returns all the necessary data representing an encoded message.

To access the encoded message, one could write:

  // A domain object |foo|
  Foo foo{...};
  auto encoded = foo.Internal__Encode();
  fidl::OutgoingMessage message = encoded.message();
  assert(message.ok());
  // Write/send the message...

This is probably pretty close in shape to the API we'd like long term,
but the return value for now is an internal type to indicate it's still
experimental, as does the function naming.

Original-Bug: 82189
Test: fx test //src/lib/fidl/cpp

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605335
Original-Revision: 091889973a8eda5f5052cfbdae30a51096f9c320
GitOrigin-RevId: 21c8438b9dc57e104d09886e1f92bedf06007073
Change-Id: I8020a20e1fbdbb6b681bd8bab631587f632ec6ce