[roll] Roll fuchsia [driver_host] Fix flake
Removing the vnode seems to result in removing a reference from the vfs
to the device, resulting in a use after free when the vfs is unable to
finish cleaning up access to the device prior to the device being freed.
The UnbindDevice call earlier in the test will result in a call to
DriverHostContext::DriverManagerRemove which will reset the device vnode
at the appropriate time.

Original-Bug: 86267
Multiply: driver-host-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/614742
Original-Revision: d0dbd5efe72ead5586f2fcefee4a8fd62a4e96c8
GitOrigin-RevId: 487ac051ee39e8873528692a2c2d64991a2f2351
Change-Id: Ia4a6a6cf73928b1bbf798ee4fbda62159cba3550