[roll] Roll fuchsia [rust][time] Migrate ClockUpdate to zx_clock_update_args_v2_t.
With v2 there are more fields and we don't want to support all
combinations. The typestate builder pattern works really nicely for this
since it ensures code that sets invalid combinations (or even sets the
same property twice) will fail to compile.

This means going to a `ClockUpdateBuilder` that is generic over the
state of each property and which may be converted to an immutable
`ClockUpdate`. This requires a small change in all callers.

Note that the advantage of zx_clock_update_args_v2_t is the ability to
specify an absolute time, but this CL doesn't take advantage of that
ability yet, all clients are converted to use "approximate_value" which
is equivalent to the v1 functionality. We temporarility allow the
combination of approximate time and rate during the transition but
intend to disallow this in the future.

Original-Bug: 65726
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/613661
Original-Revision: cfe7bdfa432a7eb8dab1428e847503cba05a657e
GitOrigin-RevId: 1cfb44517ec05e9c4eb0bd053a77f51bf2153347
Change-Id: I3cb605a83b94eec5aada956caa27682465daef05