[roll] Roll third_party/llvm-libc/src [libc] update tidy rules to fix variable formatting
This commit changes the clang-tidy rules for LLVM-libc to follow the new
format. The next commit applies these rules to the codebase.

The rules are as follows:

CamelCase for classes
lower_case for variables
lower_case for functions
UPPER_CASE for constexpr variables

There are also some exceptions, but the most important one is that any
function or variable that starts with an underscore is exempt from the
formatting.

Reviewed By: sivachandra, lntue

Differential Revision: https://reviews.llvm.org/D114301

GitOrigin-RevId: 45d7102e432b5deece5f5c6bc46576efce8992f6
Original-Revision: 0bf229abcc191632d037cf5e0166cb427f9239cf
Change-Id: Ia0a15f15fe0941fbd7ba635701592249d73a59e3