[roll] Roll fuchsia [archivist] More informative error type
Refactors all usage of `anyhow::Error` that isn't for tests into its own
error type. Additionally, the inspect collector used to fail if a single
file in the directory failed to be read. This shouldn't be the case and
this CL aligns the behavior of reading from a file with the behavior of
reading from a protocol.

Original-Fixed: 51237

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/614941
Original-Revision: dd96c8114f3762743f3b4b1c2f8c53a3b5903384
GitOrigin-RevId: 3c360f3320e863e771621da02dfd4e1e06ae46bf
Change-Id: Ifd0eb52db6c1184672cffae713aeb663fff20cec