[roll] Roll fuchsia [hub] Remove 'deleting' directory
This CL re-lands the changes introduced in fxr/605042 by only including
the minimal change set needed to remove the `deleting` directory.

The `deleting` directory is removed from struct `Instance` and no longer
serves as a holding directory for destroyed instances between the
`on_destroyed_async` and `on_purged_async` event handler functions.
References of the `deleting` directory are removed from tests.

The reason for this change is because there are no current use cases of
the `deleting` directory. Furthermore, this directory uses the instance
ID of monikers which is in the process of being deprecated (x-ref
fxr/605042).

Original-Bug: 87405
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/614983
Original-Revision: cf8762204705186118d1438d0861328da4b373a3
GitOrigin-RevId: 2af12409acafac8646f7efbfe0975cd899e9cd6f
Change-Id: I281f32df79f5029ecb2ccbfefa1b86183f686069