[roll] Roll infra/recipes [cipd_util] Fail package upload in tryjobs
We should never upload CIPD packages from tryjobs, because the code used
to publish the package is generally untested and unreviewed. It's better
for such a build to fail than it is for it to quietly upload to CIPD.

In theory, tryjob service accounts should not even have permission to
upload to CIPD, but that is not always the case and it can't hurt to
have redundant checks here.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/615310
Original-Revision: 22881f082e0532a3dacbc72f455205b987688b9c
GitOrigin-RevId: 34b9f9737e59782f731f0b67a1ca07fbd690ae53
Change-Id: Idf0035d4bf1c9e4c0ba8755fec341fb45e657040