[roll] Roll fuchsia [driver_runtime] fdf_channel_call implementation.
This adapts the zircon channel call implementation and tests.

When a client calls Channel::Call, we allocate a new txid,
and write it to the message and install a new message waiter.

When the channel receives a message of a matching txid,
it will deliver the message to the waiter and signal it.

Original-Bug: 87161

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/612140
Original-Revision: 04eae27543fdd0ea913502b443183a35f0b42992
GitOrigin-RevId: 58fe9056750cb8d3d94c6b1bbb8eb234227403d3
Change-Id: I3a3238a7c829314efcabae7c24d50d50c457a4a1