[roll] Roll fuchsia [input] Adds the modifier handler to the input pipeline
The modifier handler tracks the modifier and lock states,
and applies it to `KeyboardEvent`.

The functionality is moved from out of `KeyboardBinding`,
since the processing there is really stateless, while we
want to keep tracking key states.

This removes the modifier handling duties from `KeyboardBinding`,
which is already complex enough, and relegates them to just
after `KeyboardEvent` is injected into the pipeline.

This allows us to have minimal modifier tracking tests, since
the bulk of correctness testing is already done on the components
that actually do the tracking.

The change must be made atomically with //src/session to ensure
that the modifier handling in scene manager remains unchanged.

Original-Fixed: 89637
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/613624
Original-Revision: 71cb10332e0e8457c4e22cf4a4488f18e490c08b
GitOrigin-RevId: 2b9327951105b36462ff4da7188e7463bfdcbaed
Change-Id: I8eb9aca94f641abc3202a9c4ec449a1c024d25dc