[roll] Roll fuchsia [input] Removes the now-unnecessary modifier tracking
ImeHandler used to track the modifier and lock states.  When we added
ModifierHandler, this became unnecessary, so long as we know that
ModifierHandler is installed ahead of ImeHandler. Since we should always
do that we should be justified to remove the redundant state tracking.

This change removes the redundant state tracking and adapts the tests
accordingly.  While this is mainly a refactoring change, some tests did
not uphold the modifiers invariant on the input events, so that had to
be modified.

The invariant in question is that at least one modifier bit must be set
on a key event that reports a Shift pressed event.

Original-Fixed: 89763
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/613782
Original-Revision: 74c7e08524f2c46e12e0040d4287d98b3da2716c
GitOrigin-RevId: e251941f7455ba6000a45153d2720ba94ba6697a
Change-Id: Ic2fc5f4b51f81e6501c0a295fb4b10f971d72169