[roll] Roll fuchsia [fvm] Convert partition limits to slices.
Previously the FVM FIDL API used bytes for specifying the maximum sizes
of partitions (the "partition limit"). This was inconsistent with the
rest of the FVM API which uses slices.

This converts the partition limit to slice units and updates the users.
The fshost configuration is left in bytes since the slice size isn't
necessarily available at that layer. fshost converts to slices when it
sets the limit.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/606092
Original-Revision: 2d1f7796030c0d7fd2df6d72c8274936552d7e5b
GitOrigin-RevId: 32089032d7791e2dd9098fa5ada40738b0651280
Change-Id: I6227a7fcdf021b300c19c075dea47a320bfe349a