[roll] Roll fuchsia [sl4f][modular] Resolve session component before using its protocols
The facade connects to session protocols through a hub directory
that is created when the component is launched. However, the `exec`
subdirectory is not created until the component is *started*, leading
to a race.

Now the facade will ensure the component is resolved, and then connect
to protocols under the `resolved` directory.

Original-Bug: 87456
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/614992
Original-Revision: 7016890e9341f92f9dddd9247eec9c28c2170118
GitOrigin-RevId: 3f2f2e8b61f6ec9dba3c367da2228a1acc6e0c10
Change-Id: I67ee5ec2dc5ce16198e3c87f913b2bd2524eac79