[roll] Roll fuchsia [run-test-suite] stdout reporter part 4: update run-test-suite interface
This is part 4 in a refactor to move stdout output to a reporter
implementation. This makes a small update to the run-test-suite
interface so that callers are responsible for composing reporters and
passing them into run-test-suite.

Testing: existing integration tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/613632
Original-Revision: aaaa098e9bfc54bb625c5de350f21885e88a141f
GitOrigin-RevId: 1c533e2ac7a5b765effc62a0fbf1dc0fe57c50df
Change-Id: I589c2a8c2571e745b59b11c32ba53cac83452b2e