[roll] Roll fuchsia [component][fuzzing] Add Corpus::Load
This CL adds a way to load corpora from package storage. This allows
fuzzer authors to include their seed corpora in the fuzzing package and
supply the package-relative location as arguments to the target adapter.

The actual passing of those parameters from the target adapter to the
engine will be handled in the next CL.

Original-Bug: 24740
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/609060
Original-Revision: 139c04c3195020ea7ca21cd7ec77fcf38ff2ee82
GitOrigin-RevId: 16b3439b0bb2dccccd74d2e771605ecf05d04522
Change-Id: I07ec0713aee674a067830ecf1886f5f9bebd6c21