[roll] Roll fuchsia [fdf] Update driver references when we unbind
When a driver component is unbound, we need to update the node it is
bound to, so that the node will no longer wait for the driver to be
unbound during its own teardown. Without that, there is no way to
introspect the ServerBindingRef to be aware of the fact that the
AsyncBinding was torn down.

Multiply: interop_test
Original-Fixed: 89937
Original-Bug: 33183
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/615466
Original-Revision: 393e1c16934b9f49ae602daf3075d4ee09e2090d
GitOrigin-RevId: c7bc075a27ab9b55ab1d03ce6be70b80e3cf80db
Change-Id: Ia53c39454708e8d53ca3b6b7e824667952b56585