[roll] Roll fuchsia [fxl] Shuffle googletest cases
This causes googletest to emit the value of the random seed on each
iteration, which is nice considering we have several tests that
currently use this (and are probably all getting a random seed of zero).

Standardize retrieval of this value while I'm here.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/616181
Original-Revision: 2713ab7c0e138980acd4697d75bc865182e1cddc
GitOrigin-RevId: ade999255c9bf9a63f78cbd945e0a9c75a052400
Change-Id: I2489060a11d44c4c6b2422021e7300ee73edf768