[roll] Roll fuchsia [zircon][vm][ppb] Replace with loaned on move to active
We replace non-loaned pages with loaned pages as we put
a page in it's designated queue, if the designated queue
is an active queue.  Under normal conditions this
gradually replaces non-loaned pages with loaned pages.
Near OOM, as we're peeking pretty much all the pages
except active ones, we'll replace any pages we're moving
into the active queues.

This doesn't quite completely guarantee that we'll be
using all available loaned pages by the time we OOM, but
this gets us a lot closer to that.  To get the rest of
the way we could do a final sweep of the active queues
after eviction near OOM is complete but before we
consider eviction near OOM done.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/614224
Original-Revision: 06d5b889e8c311c65e759c3ade0f321c644e5f97
GitOrigin-RevId: 9b2de70b218d4fab64883b0a8d9162585c576c0c
Change-Id: I137854ee7089de0c158b18cd299b11291d977f62