[roll] Roll fuchsia [testrunner] Refactor ffx testing logic into an FFXTester.
Before, we included the ffxInstance as part of the FuchsiaSSHTester and
would use it if ffx was enabled. Now we separate it into its own Tester
and include an SSHTester as part of the FFXTester to run v1 tests and
fetch data sinks (not supported by ffx yet). Once all v1 tests have
migrated to v2 and fetching data sinks is supported by ffx, we can
remove the SSHTester from the FFXTester. Since there will be more logic
that needs to be added to handle processing test outputs from ffx, it
makes more sense to create a separate FFXTester rather than add it to
the FuchsiaSSHTester.

Original-Bug: 89229
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/616632
Original-Revision: 250408a653fd7946f0711c339051d7e7ddf6cc70
GitOrigin-RevId: 51ad9c62625cd08e34107b62dc0c38dcdd8b9680
Change-Id: I83dcc380bc7478638f5550b3457b406977b90460