[roll] Roll fuchsia [realm_builder] add GetRealmDecl/ReplaceRealmDecl
The GetComponentDecl and ReplaceComponentDecl functions are only capable
of interacting with children of a realm. Interacting with the decl of
the realm itself is also an important use-case, so that users may do
things like add environments or declare collections in a realm.

Add GetRealmDecl and ReplaceRealmDecl to the realm builder FIDL to
address these use-cases.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/617141
Original-Revision: 448ad469a621d0b97b7724a92f566907c768e644
GitOrigin-RevId: 0a07e0bdd29c3e4bd3d5b5a68725b44dbff730a0
Change-Id: Ifd16e7756afec4638bf8fa6c435d31aed630102e