[roll] Roll fuchsia [component][fuzzing] Add TargetAdapterClient
This CL adds a client to fuchsia.fuzzer.TargetAdapter. Both the engine
and test-engine for the framework need to interact with the target
adapter in same manner, so it is useful to extract these interactions
from RunnerImpl and encapsulate them in their own class.

Original-Bug: 24740
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/603974
Original-Revision: 20096bdff635c50418c2db94265d83174f734162
GitOrigin-RevId: 904c3a9f29a74e0f9595cb0f11c00cc57cf70c02
Change-Id: I4ac8e29418248a8954aef09dacf5d05d04fdb3b6