[roll] Roll infra/recipes [recipe_testing] Set recipe_testing props on BB subbuilds
It's important to set "$fuchsia/recipe_testing.enabled" so that builds
launched in recipes CQ don't exit early when they see that the build
graph is unaffected. I'm actually not sure why builds haven't been
exiting early in recipes CQ even when we weren't setting this property,
but in any case it can't hurt to have multiple levels of protection
against that.

Also set `recipe_depth` on Buildbucket builds as well as all led jobs.
It will be ignored on most led jobs because most of them are not recipes
CQ orchestrators, but it makes the code simpler if we set it
unconditionally.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/617184
Original-Revision: 3283967fa2ef767a79996abafe2afbc3eb83cc6e
GitOrigin-RevId: e9431168c006c58f162af37129a44e9341558223
Change-Id: Id26278da37981b2223de208802f8d8d634a6b69f