[roll] Roll fuchsia [time] Move tests to absolute time where beneficial.
In a couple of cases the fact that a test couldn't directly control when
a clock update occured led to a less accurate test. This fixes those
without changing the logic of how time is maintained.

Note than in many cases a test just needed to start a clock or set a
coarse time, and in those cases approximate_time is fine.

Original-Bug: 65726
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/617782
Original-Revision: 1ec7641580fc74c43abcd5cca0d67efcf7d12d7a
GitOrigin-RevId: 010ead8fa9acefc205cca56b9c3e761678418e78
Change-Id: I401f5f6ca83c72b93c616fd9e65a99e6af7c03cb