[roll] Roll fuchsia [lib][concurrent] Add libconcurrent to the kernel. (RE-LAND)
This CL is a re-land of http://fxr/606085 which was reverted in
http://fxr/616782 because of http://fxb/90171

Please refer to the bug for details.  The original commit message was
as follows.

"""
Add libconcurrent to the kernel as kconcurrent, defining lockdep
guards in the process.  Also add unittests for the guards.
"""

Original-Fixed: 90171
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/618782
Original-Revision: bbcf0326a935917f466f5f3052b2b5b22df3eeec
GitOrigin-RevId: e45926d6f803af438af2669f32bcb64d77464562
Change-Id: I548ac38fcd865ee8a62cc3949190fc7635e676fe