[roll] Roll fuchsia [fdf] Stop the driver component on Stop or Kill
When either ComponentController.Stop() or ComponentController.Kill() are
called, make sure we stop the driver component.

Previously, there was an ordering guarantee that Kill() would be called
after Stop(), and that at least Stop() would be issued. That is no
longer true, and though we set a stop-timeout for the environment that
the driver components are run in, this provides some additional caution.

This change also ensures that we correctly send an epitaph to the
component manager, so knows the component runner has stopped the
component.

Original-Bug: 33183
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/618005
Original-Revision: 4c42401271e1fecca5718f3c62f14a857d6faa1a
GitOrigin-RevId: 6655295b6f082db7ec5a589d8f2abf7f6473cab6
Change-Id: I9317474cb261d3ee0afc276394f6d2f5f327de42