[roll] Roll fuchsia [wlan][policy] Add connection duration roam metric
Log a metric for how long the device was connected to an AP before
roaming to another AP or how long it was connected before an unexpected
disconnect and the total of both to compare.

Currently the decision of whether or not a roam should happen is not
implemented, but when it is implemented it will send a telemetry event
for the proactive network switch.

Test: added unit tests: fx test wlancfg-test,
      manually tested with also_log_locally for non-roam disconnectss
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/616661
Original-Revision: de58ad838e66893df2695deeff334ea310a354eb
GitOrigin-RevId: ead8e8c6a5f6bab77a34bbdaa655b1e23f6d2c98
Change-Id: Id7dc03718fb3372409b0e79c2b5d932b99f576cf