[roll] Roll fuchsia [debugger] Handle stripped ELF symbols.
Stripped ELF binaries are missing the .symtab section. In this case the
only ELF symbols we can use are in the .dynsym section.

This merges the symbols from both .symtab and .dynsym when reading ELF
symbols so we don't miss any, regardless of the stripped status of the
binary.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/619081
Original-Revision: 4c5942b18f96145352ccda5e40456dbc08333056
GitOrigin-RevId: 97336d0f20a863537e9cbd19e87172d48461648e
Change-Id: Ib39ed0e258e74f23c22bd82d9a232b48efccb7ec