[roll] Roll fuchsia [kernel][hypervisor][arm64] Do not convert paddr
The pmm arena base is already a paddr_t and the previous code that
treated it as a vaddr and attempted to convert it to a paddr was
incorrect.

This bug was benign for so long since the relevant ARM boards
that virtualization runs on have their physical memory starting at 0,
and all of this memory would be presented to the kernel as one large
arena. Although the conversion of vaddr 0 to a paddr would fail, it
would fail by giving a value of 0, which coincidentally was the right
value.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/619302
Original-Revision: cca9b3316f8271326a0e38ddccd71fbfa8ea800e
GitOrigin-RevId: 653e15b750bf63a7f3648f82ff47492d604ebd7d
Change-Id: I48ffd8b309bf017fded3eb6df85d722eb6d515d8