[roll] Roll fuchsia [netstack3] Check IdMap contents in property test
The existing property test checks that arbitrary sequences of operations
don't leave the contents of the IdMap in an invalid state. This patch
extends that by adding a HashMap that represents "a source of truth". By
applying the same operations to both the IdMap and HashMap, we can
verify that the IdMap implements its API faithfully. As a side-effect,
this cleans up the `let _ = map.Xxx()` calls in Operation::apply since
the results of applying operations on the map are now used instead of
ignored.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/615661
Original-Revision: 00cbbb22d57e9f245736fb16ccd8737df3732786
GitOrigin-RevId: 5e639a921d8f8b3ee31b7422b677bbefba7ee47c
Change-Id: I46a831b4dbe3923def911d6c7bf60dadfb07358f