[roll] Roll fuchsia [test][kernel_panic] Add SMEP/PXN tests.
This CL adds an emulator test that checks that the kernel crashes when
it tries to execute user memory.

A new k command is added: `k crash_user_execute` creates an
userspace mapping, copies a simple function into it, remaps it as
read-only executable and calls that code.

The test should cause a kernel crash in both arm64 (PXN) and x86-64 (SMEP).

This CL also renames the SMAP test so it is architecture agnostic, to
avoid confusions.

Test: fx test kernel_panic_tests
Original-Bug: 88451
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/611946
Original-Revision: 194eb2de832e80318255f258a6cb3162d73b446b
GitOrigin-RevId: f8452ae590c1da225e7cfa08c2aa913e94c89d7a
Change-Id: Iabf041f137284febde43f26e2ee9ed020764d390