[roll] Roll fuchsia [storage] Allow connections to services on R/O filesystems
Connections to services are usually made with read and write rights
(fdio_service_connect does this), but these would always be rejected if
a filesystem was marked as read-only.  This change allows connection to
services even if the filesystem is read-only.  The parent connection
rights should still be honored, so write access would still be required
to the parent connection.

Original-Bug: 89869
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/617241
Original-Revision: dd084a633babf6bfebdb987f6202158aea2a388d
GitOrigin-RevId: c4b4c82e19632949257bffcc48bddcbc5e273b74
Change-Id: Ida6b7006057f7c02f8295b03baa24bae1020d890