[roll] Roll fuchsia [timekeeper] Don't automatically start monitor clock.
We're not currently using a monitor clock, but the code to create it
currently follows the old primary clock behavior where component
manager used to start the clock before giving it to us. Its more
consistent to create a clock that hasn't started and let the
clock_manager take care of the start when its ready.

Original-Bug: 65726
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/617710
Original-Revision: dbf8aa10dbdea99986ddcb1093b67ba874bfb421
GitOrigin-RevId: 9416e2e648dc3f9a63752314b036f37cfb400eee
Change-Id: Iee9cb47582b2187a6243b0022faa7bce18b3e809