[roll] Roll fuchsia [fit] Replace bind_member with more-efficient variant
This CL deprecates the existing version of fit::bind_member,
which uses the fit::bind_member(&instance, &Obj::Method) syntax and
introduces a new variant with a different API
fit::bind_member<&Obj::Method>(&instance).

The motivation for this is that the existing fit::bind_member is
inefficient and the lambda is 24-bytes, leading to a heap allocation
in a fit::function. In contrast, the new fit::bind_member's lambda
only requires 8-bytes.

The idea is to migrate existing usages to the new fit::bind_member
where possible (C++17 and up) and migrate to a regular lambda in
other cases (C++14 and down).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/612883
Original-Revision: 7d7cba3b1d0b07451142b3a9104b2204fdbf250b
GitOrigin-RevId: 96e1b8906d669edeb8987ab43a777bf78b0f167e
Change-Id: I999ed4ad32fad800ed9c58770be417e219fa72ff