[roll] Roll infra/recipes [snap_branch] Don't short-circuit petal snap for fast-forwards
It is possible for the post-snap git revision to be equal to the target
petal revision (the revision tracked in integration.git), but not be
equal to the branch's remote revision, in the case of a branch
fast-forward. We were short-circuiting this case, which would cause the
petal branch to go out-of-sync with integration.git.

This CL changes the logic to short-circuit more strictly, i.e. the
post-snap revision, target petal revision, and remote branch revision
all have to be consistent (the desired state) in order to determine that
there is no work to do.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/619941
Original-Revision: 23908189003e2e0812f3662d8c30c18fa12e07cb
GitOrigin-RevId: b54cb38bb51ef619e3b92b8369b48ac0a698d352
Change-Id: Ifa0d0bdca82e1f445d9576679527e993ea498253