[roll] Roll fuchsia [llcpp_drivers] Generate skeleton driver code
This CL adds *_driver_wire_messaging.{h,cc} files and generates output
in those files for protocols marked with @transport("Driver"). For now,
the output is specialized for channels, but over time it will be updated
to be specialized for drivers instead.

In addition, this CL adds golden file support and introduces two goldens
- one for one way calls and one for two way calls. The one way call test
also covers domain objects so it is possible to ensure that domain
objects are reused correctly.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/617264
Original-Revision: 66b2794436ed6ed09a40e92ed2deb4bbb4b45061
GitOrigin-RevId: 7429c0b2b66dbc672b3533f05473f431a9c19d41
Change-Id: Iac74f89f8c1a99445ca960aacba34482d4061cab