[roll] Roll experiences [login] Wire login UX with AccountManager
This change wires up the password authentication logic with the login
UX. It moves the authentication logic from workstation_session to login
shell. The login shell implements the create account with password
functionality and subsequently login in using that password. The ermine
user shell implements logout using a keyboard shortcut.

The ability to mount the encrypted partition as "account_data" directory
is still blocked on http://fxb/89690 and fixes to flutter runner.

Although everything is wired up and functional (except for the flutter
runner fix), it is behind a build flag 'start_oobe', which is set to
false. In addition, the code that enables this is also hard disabled
until we are ready to land the user data protection feature.

Original-Fixed: 85564,85587,87678,87679
Original-Bug: 85576

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/614005
Original-Revision: 3c21f000c490e923a5f71d626392a68d56385b72
GitOrigin-RevId: 8883a126a82f179aa2719aa24bbf60479e8a68fc
Change-Id: Ia825f385fd3281fcdbb115a14e14eca1cde3481e