[roll] Roll fuchsia [dhcpv6] Extract address assignment test setup fn
Introduce a test setup fn to create a client in stateful mode, and
perform message exchange to assign the configured addresses to the
client. The functionality of the test setup fn is extracted from the
`assign_addresses` test, with the addition of making the addresses to be
assigned configurable. The test fn is intended to be used as test setup
for future state machine testing starting in `AddressAssigned` state.

Add an utilitary Identity Associsation test type to configure the
addresses to be assigned, specifying T1/T2 values; intended to be used
in future Renew/Rebind testing as well.

Original-Bug: 88838

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/604122
Original-Revision: 4affc12744c06cdceedeab2412aad6a38bc8f4f8
GitOrigin-RevId: f1b6b23c1d5551c759032ca71a873bd5b86e3cac
Change-Id: Idfb4ec33f5f5de225479d92b1098d9f51b58b762