[roll] Roll infra/recipes [build] Copy assembly inputs instead of symlinking
The cipd client does not follow symlinks, so if given a tree of symlinks
to files outside the package root it will simply upload a package full
of invalid links.

Ideally we would hardlink instead, but there isn't an existing solution
in recipes for hardlinking a tree of files. So in the ineterest of
fixing this quickly I'm using the existing
`api.cas_util.hardlink_tree()` function – which, despite the name,
actually copies files under the hood instead of hardlinking to work
around a CAS client quirk.

It's definitely a hack to use this CAS-specific function for uploading
to CIPD but it's by far the fastest way to get it working.

Original-Bug: 87010
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/620224
Original-Revision: 57a1571dfbf96bb06d70ac7ce5c09b5ce32b13b1
GitOrigin-RevId: 7ef06ab91b1ce4a0dc1bb90f7dd90a233d1918cd
Change-Id: If7d7dddfe0fb19d8f725745cbe4b89af4d91aef0