[roll] Roll fuchsia [llcpp] Update LLCPP examples to not heap allocate w/ bind_member
The currently used fit::bind_member is deprecated because it heap
allocates when used in a fit::function. This usage is replaced with
a lambda. The new, more efficient, fit::bind_member cannot be used
because it doesn't work with method overloads.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/619825
Original-Revision: fbaa4f96794133be1a4078d0edc5129021f867e9
GitOrigin-RevId: 2e016c6d9109aabde329269f15e51d03ff0cbd12
Change-Id: Ia8a124eb292e25395e42f705516ec0c64f19ef66