[roll] Roll fuchsia [weave] Mark success on duplicate removals
The WARM state machine stops if we fail to remove an address when
requested, but this can happen in situations where the interface was
already removed in between calls to WARM. In these cases, produce a
warning log but allow the state machine to progress. Tests are added to
confirm behaviors on duplicate add and remove.

Original-Bug: 89993
Test: fx test weavestack-adaptation-unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620006
Original-Revision: 0ebfef1f35a0c2a4d7e32ba05ffcf51863db24db
GitOrigin-RevId: ff3e68c08abc45163af6188bce2576b1b2ed0e04
Change-Id: I7e47d18dbe61a3adfce86cbb9917e5dd31c285f7