[roll] Roll fuchsia [ffx] Remove previous Behavior code from ffx emu
We're going a different direction with the idea of Behaviors in ffx emu.
We'll be utilizing HandleBars to simplify the specification and code to
handle it.

This CL prepares for the conversion by removing all of the previous
iteration of Behaviors from the plugin's code, as well as the PBM code.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620102
Original-Revision: cb4056870af17ca398d0db5f2f334b2026416758
GitOrigin-RevId: 3ba98d30fcdcac3fe0ff8680b9f1bf8074d95586
Change-Id: I3509848d6abbba74de2105ef417cf3c990d6da49