[roll] Roll fuchsia [build] Fix variant selection in fuzzer_package() template.
A small CL to fix the way variant selection is performed in the
fuzzer_package() template. The old code could only work with
very restricted selectors definitions, and would not work
with more complex ones, for example:

  {
    variant = "asan-fuzzer"
    output_name = [
      "first_fuzzer",
      "second_fuzzer",
    ]
  }

Because `output_name` is a list with two items here.
Or even:

  {
    variant = "ubsan-fuzzer"
    dir = [ "//src/foo" ]
  }

Because the `dir` was not supported.

+ Update a comment in //build/zircon/BUILD.gn to point to
  the right .gni file.

Original-Bug: NONE
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/619644
Original-Revision: ee4ac47177540fabdf760188baa1a5bcf70ba850
GitOrigin-RevId: 3398f8118a29fa64ea4e97c32db58cd0e0f36c15
Change-Id: Ifc80cd229be3926d474f70282d5e7a188dd2209b