[roll] Roll fuchsia [input] Add event counter breakdown by event type
This might help us understand what is going on with intermittent keyboard issues.
See attached bug for details.

Changes are as follows:

- added separate event counters for each event type we know of
- added event counters for handled events for each event type
- added appropriate tests

The hope here is that detailed event breakdown helps us identify when the
key events are not being received from the driver, versus being lost in
the input pipeline, or in the consumers.

Original-Bug: 90109
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620522
Original-Revision: 76935bc1f67c668f700401b551bea353f96d10ab
GitOrigin-RevId: f5f6b89e4f609acec1f95ab5d0bfcfb9bb4c8e15
Change-Id: Ice0f4d3dc384a81ec56fe236ce08e6eb0a0da400