[roll] Roll cobalt Local aggregation only cleans up data after observations are stored.
I discovered this by creating a realistic integration test for the local
aggregation. This expands an existing local aggregation test to be more
representative of how the code is called in practice, and to cover more
scenarios.

It also exposed an issue with the SELECT_MOST_COMMON aggregation
procedure, that it was creating empty observations, which is fixed
as well.

Original-Bug: 89902

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/614901
Original-Revision: ab014a68999c42d79eb205edec0fc45395c2fb16
GitOrigin-RevId: 8b8f08f297f176d5a39297dbd6877d3fa154098f
Change-Id: I6066a333bd49b6e3044c66999a66778795fe1ace