[roll] Roll fuchsia [fx] Make manifest mismatch error actionable
This most often happens when a developer with a fully internal checkout
tries to reproduce an internal build that used the internal integration
repo, but with a public manifest.

In that case the public manifest contains a subset of the internal
manifest that the developer is likely using locally, so syncing to the
revision used by the build is probably good enough.

However, this is not sufficient for CQ builds (where "patches" is
non-empty) because the manifests need to match in order for jiri to
patch in the change used by the build.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620781
Original-Revision: 5b746b7b42eaa776e934703a9c160e471e8994e0
GitOrigin-RevId: b69b61359647fcbcfaf1e0a21de672bf22fe1288
Change-Id: Ie3a629f462a91e4296b27a7d68119146c247422a