[roll] Roll fuchsia [kernel][vm] Fit lambda within kDefaultInlineCallableSize
Reduce the capture size of a lambda used in
PagerDispatcher::QueryDirtyRanges to fit within
kDefaultInlineCallableSize which is 16 bytes. Previously the lambda was
greater than 16 bytes but was getting the correct behavior when used to
initialize an inline fbl::Function sized 32 bytes due to alignment
(16 bytes for the capture, 8 for the function pointer and the remaining
for alignment). Do not rely on alignment and instead size the capture
within kDefaultInlineCallableSize as expected.

Original-Bug: 63989
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620889
Original-Revision: 5d0d28f477aa1de1886f3cedfa21a887ad943ce8
GitOrigin-RevId: f98b576771aebd27f9af2c60483e328b61fe3193
Change-Id: I5a5252acfef62311e1ebe6afd4c56f92294c71cc