[roll] Roll fuchsia [ffx] Move command line flags to emulator config
This CL adds a FlagData type and field to the emulator configuration.
This is similar to the previous Behaviors map, except it's all in one
structure instead of a map of them. There are still the same four
fields: args, features, kernel_args, and options. These map to the parts
of the command line they represent.

We also remove extraneous fields from the engines, and adjust the
build_emu_cmd accordingly.

Original-Bug: fxbug.dev/90481

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620104
Original-Revision: 9361132be022828e262aed56ffab97c7882a7d45
GitOrigin-RevId: 0d2b60c1497d9b987e7a8c52239258fd17d544ec
Change-Id: Ia3bba8426647671d7e61581ac4eedece846a166d