[roll] Roll fuchsia [run-test-suite] Add back test name preamble
testparser relies on this preamble to exist to recognize the test as
a test run using test framework. Removing this broke testparser and
it could no longer collect information on test cases.
While we're moving towards a world in which tools have a structured
output to parse, the migration isn't complete and we shouldn't break
tools in the meantime.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621145
Original-Revision: a38092e25d20b11cc72a2faebcebc40a225febc1
GitOrigin-RevId: 990dbe25f67a5397c31e051cacea18debf69e315
Change-Id: I54a45dc59516adf6fddb618c617f7ed621ee499b