[roll] Roll fuchsia [virtualization][linux_runnuer] Automatic setup of GPU drivers
Don't rely on prebuilt container for this. This makes our GPU
drivers work with a fresh container.

This is also adding the magma device and setting up the drivers
before launching the container shell in order for the initial
Xwayland instances to use the correct drivers.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621182
Original-Revision: 130a28308b00cbc0cfdac086060d3bdc6381d1a9
GitOrigin-RevId: 7d920f829746d863081cf3cf9c590e8ca97d1dfc
Change-Id: Id8385d18b135a2540dc32683b06c37ecd4788df8