[roll] Roll fuchsia [virtualization] Use an FVM partition for the stateful image
Update the linux_runner so that when we use a stateful image, we will
look for an FVM partition to back the virtio-block device that is used
for the stateful image. If no such FVM partition exits, than we will
create it.

Original-Bug: 90276
Original-Bug: 90469
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/614422
Original-Revision: 331a836b677cae9074dadc7000a1a0fde0e06b74
GitOrigin-RevId: d67b172f8663a58433b2912a29940813fc1233b2
Change-Id: I82cf1a5f849921c5f030823a22277904540d1a40